hotfix: check user id before removing topic

+++ __init__.py
@@ -190,6 +190,11 @@ def deletePlace():
 
         place_to_delete = session.query(Place).filter_by(category_id=category_id).filter_by(name=name).one()
 
+        if user_id != place_to_delete.user_id:
+            return "<script>function myFunction() {\
+                    alert('You are not authorized to delete this place.');\
+                }</script><body onload='myFunction()'>"
+
         session.delete(place_to_delete)
         session.commit()
         return redirect(url_for('showMainPage'))
